Moved selection commands inside `editor.js` (#220)
* Moved commands

* added selection command

* attempt to fix the flaky test in dynamic content other test

* flaky test fix: added waitUntil

* fix test
