+++ cypress/support/stackable/commands/toolbar.js
@@ -66,16 +66,20 @@ export function adjustDynamicContent( blockName, blockSelector, selector, option
 			.type( '{selectall}', { force: true } )
 	} else if ( typeof selector === 'function' ) {
 		selector()
-		cy.wait( 1000 )
 	}
 
 	cy.adjustToolbar( 'Dynamic Fields', () => {
-		const selectFromSuggestions = option => cy
-			.get( '.stackable-dynamic-content__popover-content' )
-			.contains( containsRegExp( option ) )
-			.parentsUntil( '.components-base-control' )
-			.find( '.stackable-dynamic-content__input-container>input' )
-			.click( { force: true } )
+		const selectFromSuggestions = ( option, value = '' ) => {
+			cy
+				.get( '.stackable-dynamic-content__popover-content' )
+				.contains( containsRegExp( option ) )
+				.parentsUntil( '.components-base-control' )
+				.find( '.stackable-dynamic-content__input-container>input' )
+				.click( { force: true } )
+				.type( `{selectall}${ value }` )
+
+			cy.waitLoader( '.components-spinner' )
+		}
 
 		const selectOption = option => cy
 			.get( '.react-autosuggest__suggestions-container--open' )
@@ -89,14 +93,14 @@ export function adjustDynamicContent( blockName, blockSelector, selector, option
 
 		if ( Array( 'Other Posts', 'Latest Post' ).includes( source ) && post.length ) {
 			// Select a post if source is Other Posts / Latest Post
-			selectFromSuggestions( `${ source === 'Other Posts' ? 'Posts/Pages' : 'Nth Latest Post' }` )
+			selectFromSuggestions( `${ source === 'Other Posts' ? 'Posts/Pages' : 'Nth Latest Post' }`, post )
 			cy
 				.get( '.react-autosuggest__suggestions-container--open' )
 				.contains( post )
 				.click( { force: true } )
 		}
 
-		selectFromSuggestions( 'Field' )
+		selectFromSuggestions( 'Field', fieldName )
 		selectOption( fieldName )
 
 		if ( ! isEmpty( fieldOptions ) ) {
